Update support range for koto
